pyproject.toml
[project]
name = "txtosqlbot"
version = "0.1.0"
description = "AI agent that lets you to talk to your database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.21",
    "langchain-groq>=0.3.1",
    "langchain-ollama>=0.3.0",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "rich>=13.9.4",
    "streamlit>=1.44.0",
    "watchdog>=6.0.0",
    "groq>=0.4.1",
]

[dependency-groups]
dev = [
    "faker>=37.1.0",
    "pre-commit>=4.0.1",
    "ruff>=0.8.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.setuptools]
packages = ["txtosqlbot"]
